- Switched to build with GCC 4.2 - Enabled NS_BUILD_32_LIKE_64 to make CG/NS structs the same and help improve   type checking for 64 bit - Cleaned up a bunch of int/unsigned int types that need to be NSInteger or   NSUInteger.  More to come. - Removed OWF's caching support as it is no longer used by OWF or OmniWeb - Added support in OBPostLoader for finding and warning about classes that add   object-typed ivars, whose superclass conform to NSCopying, but don't   themselves implement -copyWithZone:.  These classes often need to implement   their own -copyWithZone: to either retain, copy or clear the copy's ivars.   There are false positives, but this is a helpful checklist. - Added a bunch of missing -copyWithZone: methods. - Added -cgFloatValue on NSControl and OFNaN, mirroring the version on   NSNumber and NSString. - Cleaned up code using fast enumeration, @properties, C99. - Split guts of OBAssertFailed into OBInvokeAssertionFailureHandler so that   they can be used independently (OBAssertFailed services as the debugger   breakpoint target only now). - Added OBASSERT_NOT_IMPLEMENTED for helping track down stray implementations   of API that will no longer be called on subclasses or delegates. - Added support for fetching the committed row count matching a predicate on   ODODatabase. - OFBinding how supports key paths throughout for both source and destination   keys. - Fixed some date parsing issues in OFRelativeDateParser and added tests - Fixed buggy hex-formatted string->data conversion and made it use NSError   instead of raising. - Added conveniences for accessing XML attributes of various types with   default values.

+++ Frameworks/OmniAppKit/OAApplication.m
@@ -125,7 +125,7 @@ static unsigned int launchModifierFlags;
     }
 }
 
-- (void)endSheet:(NSWindow *)sheet returnCode:(int)returnCode;
+- (void)endSheet:(NSWindow *)sheet returnCode:(NSInteger)returnCode;
 {
     NSWindow *docWindow;
     OASheetRequest *queuedSheet = nil;
